Remove unnecessary static_cast in test sources

+++ tests/Demo/Source/arbitraryTensor.cpp
@@ -47,7 +47,7 @@ LIBRARY_LINK_FUNCTION(loadArray) {
 	catch (std::exception& e) {
 		err = LLErrorCode::FunctionError;
 	}
-	return static_cast<int>(err);
+	return err;
 }
 
 EXTERN_C DLLEXPORT int getElementShared(WolframLibraryData libData, mint Argc, MArgument *Args, MArgument Res) {
@@ -63,7 +63,7 @@ EXTERN_C DLLEXPORT int getElementShared(WolframLibraryData libData, mint Argc, M
 	catch (std::exception& e) {
 		err = LLErrorCode::FunctionError;
 	}
-	return static_cast<int>(err);
+	return err;
 }
 
 EXTERN_C DLLEXPORT int getElementNonShared(WolframLibraryData libData, mint Argc, MArgument *Args, MArgument Res) {
@@ -80,7 +80,7 @@ EXTERN_C DLLEXPORT int getElementNonShared(WolframLibraryData libData, mint Argc
 	catch (std::exception& e) {
 		err = LLErrorCode::FunctionError;
 	}
-	return static_cast<int>(err);
+	return err;
 }
